% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{beta.test}
\alias{beta.test}
\title{PERMANOVA for phyloseq}
\usage{
beta.test(physeq, group, distance)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing merged information of abundance,
taxonomic assignment, sample data including the measured variables and categorical information
of the samples, and / or phylogenetic tree if available.}

\item{group}{(Required). Character string specifying name of a categorical variable that is preferred for grouping the information.
information.}

\item{distance.}{A string character specifying dissimilarity index to be used in calculating pairwise distances (Default index is "bray".).
"unifrac","wunifrac","manhattan", "euclidean", "canberra", "bray", "kulczynski", "jaccard", "gower", "altGower",
"morisita", "horn", "mountford", "raup" , "binomial", "chao", "cao" or "mahalanobis".}
}
\description{
PERMANOVA for phyloseq
}
\author{
Kai Guo
}
