% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{buildTree}
\alias{buildTree}
\title{contruction of plylogenetic tree (extreme slow)}
\usage{
buildTree(seqs)
}
\arguments{
\item{seqs}{DNA sequences}
}
\value{
tree object
}
\description{
contruction of plylogenetic tree (extreme slow)
}
\author{
Kai Guo
}
