% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{do_ttest}
\alias{do_ttest}
\title{do t.test}
\usage{
do_ttest(x, group, ref = NULL, ...)
}
\arguments{
\item{x}{data.frame with sample id as the column name, genes or otu as rownames}

\item{group}{group factor used for comparison}

\item{ref}{reference group}
}
\description{
do t.test
}
\author{
Kai Guo
}
