% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{phy_tree}
\alias{phy_tree}
\title{Retrieve phylogenetic tree (phylo-class) from object.}
\usage{
phy_tree(physeq, ...)
}
\arguments{
\item{physeq}{(Required). An instance of phyloseq-class that contains a phylogenetic tree.
If physeq is a phylogenetic tree (a component data class), then it is returned as-is.}

\item{...}{parameters for the phy_tree function in phyloseq package}
}
\description{
Retrieve phylogenetic tree (phylo-class) from object.
}
