% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotalpha}
\alias{plotalpha}
\title{plot alpha diversity}
\usage{
plotalpha(
  physeq,
  method = c("Simpson", "Shannon"),
  group,
  pvalue = 0.05,
  padj = NULL,
  wilcox = FALSE
)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing merged information of abundance,
taxonomic assignment, sample data including the measured variables and categorical information
of the samples, and / or phylogenetic tree if available.}

\item{method}{A list of character strings specifying \code{method} to be used to calculate for alpha diversity
in the data. Available methods are: "Observed","Chao1","ACE","Richness", "Fisher", "Simpson", "Shannon", "Evenness","InvSimpson".}

\item{group}{group (Required). A character string specifying the name of a categorical variable containing  grouping information.}

\item{pvalue}{pvalue threshold for significant dispersion results}

\item{padj}{adjust p value threshold for significant dispersion results}

\item{wilcox}{use wilcoxon test or not}
}
\value{
Returns a ggplot object. This can further be manipulated as preferred by user.
}
\description{
plot alpha diversity
}
\author{
Kai Guo
}
