% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotquality}
\alias{plotquality}
\title{plot the quality for the fastq file}
\usage{
plotquality(file, n = 5e+05, aggregate = FALSE)
}
\arguments{
\item{file}{(Required). character. File path(s) to fastq or fastq.gz file(s).}

\item{n}{(Optional). Default 500,000. The number of records to sample from the fastq file.}

\item{aggregate}{(Optional). Default FALSE. If TRUE, compute an aggregate quality profile for all fastq files provided.}
}
\value{
figure
}
\description{
plot the quality for the fastq file
}
\examples{
\dontrun{
plotquality(system.file("extdata", "sam1F.fastq.gz", package="dada2"))
}
}
