% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{prefilter}
\alias{prefilter}
\title{filter the phyloseq}
\usage{
prefilter(physeq, min = 10, perc = 0.05)
}
\arguments{
\item{physeq}{A \code{phyloseq} object containing merged information of abundance,
taxonomic assignment, sample data including the measured variables and categorical information
of the samples, and / or phylogenetic tree if available.}

\item{min}{Numeric, the threshold for mininal Phylum shown in samples}

\item{perc}{Numeric, input the percentage of samples for which to filter low counts.}
}
\value{
filter phyloseq object
}
\description{
filter the phyloseq
}
\examples{
\dontrun{
data("GlobalPatterns",package="phyloseq")
physeq<-prefilter(GlobalPatterns)
}
}
\author{
Kai Guo
}
