% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{preRef}
\alias{preRef}
\title{Download the reference database}
\usage{
preRef(ref_db, path = ".")
}
\arguments{
\item{ref_db}{the reference database}

\item{path}{path for the  database}
}
\value{
the path of the database
}
\description{
Download the reference database
}
\examples{
\donttest{
preRef(ref_db="silva",path=tempdir())
}
}
\author{
Kai Guo
}
