% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{richness}
\alias{richness}
\title{calculat the richness for the phyloseq object}
\usage{
richness(physeq, method)
}
\arguments{
\item{physeq}{phyloseq object}

\item{method}{A list of character strings specifying \code{method} to be used to calculate for alpha diversity
in the data. Available methods are: "Observed","Chao1","ACE","Richness", "Fisher", "Simpson", "Shannon", "Evenness","InvSimpson".}
}
\value{
data.frame of alpha diversity
}
\description{
calculat the richness for the phyloseq object
}
\author{
Kai Guo
}
