% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{sample_data}
\alias{sample_data}
\title{extract sample information}
\usage{
sample_data(physeq, ...)
}
\arguments{
\item{physeq}{(Required). A data.frame-class, or a phyloseq-class object.}
}
\description{
extract sample information
}
