% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{otu_table}
\alias{otu_table}
\title{extract otu table}
\usage{
otu_table(physeq, ...)
}
\arguments{
\item{physeq}{(Required). An integer matrix, otu_table-class, or phyloseq-class.}

\item{...}{params to otu_table(phyloseq function)}
}
\description{
extract otu table
}
